Cleanup colorwheel subsystem
